b4/nng/pyproject.toml
[project]
name = "nng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio-gevent>=0.2.5",
    "bottle==0.13.4",
    "docopt>=0.6.2",
    "gevent-websocket==0.10.1",
    "honcho>=2.0.0",
    "ipython>=8.37.0",
    "numpy>=2.2.6",
    "ollama>=0.5.1",
    "pgvector>=0.4.1",
    "pip>=25.1.1",
    "plotext>=5.3.2",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.9",
    "pyaml>=25.7.0",
    "uvloop>=0.21.0",
    "websocket-client>=1.8.0",
]
